Add support for searching document collections

+++ documentindex/src/main/java/org/insightcentre/nlp/saffron/documentindex/CorpusTools.java
@@ -12,6 +12,7 @@ import java.util.Enumeration;
 import java.util.HashMap;
 import java.util.Iterator;
 import java.util.NoSuchElementException;
+import java.util.Random;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipException;
 import java.util.zip.ZipFile;
@@ -32,6 +33,7 @@ import org.insightcentre.nlp.saffron.documentindex.tika.DocumentAnalyzer;
  * @author John McCrae <john@mccr.ae>
  */
 public class CorpusTools {
+
     /**
      * Create a corpus from a folder, each file will be considered a single
      * document
@@ -47,6 +49,7 @@ public class CorpusTools {
     }
 
     public static class FolderIterator implements Iterator<File> {
+
         final File[] files;
         int i = -1;
         Iterator<File> currIter = null;
@@ -68,37 +71,39 @@ public class CorpusTools {
             advance();
             return f;
         }
-        
+
         public void advance() {
             next = null;
-            if(currIter != null && currIter.hasNext()) {
+            if (currIter != null && currIter.hasNext()) {
                 next = currIter.next();
-                while(currIter != null && !currIter.hasNext() && i < files.length - 1) {
+                while (currIter != null && !currIter.hasNext() && i < files.length - 1) {
                     i++;
-                    if(files[i].isDirectory()) {
+                    if (files[i].isDirectory()) {
                         currIter = new FolderIterator(files[i].listFiles());
                     } else {
                         currIter = null;
-                    }                    
+                    }
                 }
             } else {
-                while(next == null && i < files.length) {
+                while (next == null && i < files.length) {
                     i++;
-                    if(i >= files.length) {
+                    if (i >= files.length) {
                         // end of list
-                    } else if(files[i].isDirectory()) {
-                        currIter  = new FolderIterator(files[i].listFiles());
-                        if(currIter.hasNext())
+                    } else if (files[i].isDirectory()) {
+                        currIter = new FolderIterator(files[i].listFiles());
+                        if (currIter.hasNext()) {
                             next = currIter.next();
+                        }
                     } else {
                         next = files[i];
                     }
                 }
             }
         }
-        
+
     }
     private static final Document.Loader TIKA_LOADER = new DocumentAnalyzer();
+
     private static class FolderCorpus implements Corpus {
 
         private final File folder;
@@ -145,18 +150,35 @@ public class CorpusTools {
      * @return A corpus object
      */
     public static Corpus fromZIP(File zipFile) {
+        return fromZIP(zipFile, null);
+    }
+
+    /**
+     * Create a corpus from a zip file, each file will be considered a single
+     * document
+     *
+     * @param zipFile The zip file
+     * @param targetDir The directory to extract the file to
+     * @return A corpus object
+     */
+    public static Corpus fromZIP(File zipFile, File targetDir) {
         if (!zipFile.exists() && zipFile.isDirectory()) {
             throw new IllegalArgumentException(zipFile.getName() + " does not exist or is a folder");
         }
-        return new ZIPCorpus(zipFile);
+        if (targetDir != null && !targetDir.mkdirs()) {
+            throw new IllegalArgumentException(targetDir.getName() + " could not be created as a file");
+        }
+        return new ZIPCorpus(zipFile, targetDir);
     }
 
     private static class ZIPCorpus implements Corpus {
 
         private final File zipFile;
+        private final File targetDir;
 
-        public ZIPCorpus(File zipFile) {
+        public ZIPCorpus(File zipFile, File targetFile) {
             this.zipFile = zipFile;
+            this.targetDir = targetFile;
         }
 
         @Override
@@ -184,11 +206,18 @@ public class CorpusTools {
                                     if (ze.isDirectory()) {
                                         return null;
                                     }
-                                    if (file != null) {
+                                    if (file != null && targetDir == null) {
                                         file.delete();
                                     }
-                                    file = File.createTempFile(ze.getName(), "");
-                                    file.deleteOnExit();
+                                    if (targetDir != null) {
+                                        file = new File(targetDir, ze.getName());
+                                        while (file.exists()) {
+                                            file = new File(targetDir, ze.getName() + new Random().nextInt(10000));
+                                        }
+                                    } else {
+                                        file = File.createTempFile(ze.getName(), "");
+                                        file.deleteOnExit();
+                                    }
                                     FileOutputStream fos2 = new FileOutputStream(file);
                                     InputStream is = zip.getInputStream(ze);
 
@@ -219,26 +248,43 @@ public class CorpusTools {
         }
     }
 
-     /**
-     * Create a corpus from a tarball (.tar.gz) file, each file will be considered a single
-     * document
+    /**
+     * Create a corpus from a tarball (.tar.gz) file, each file will be
+     * considered a single document
      *
      * @param zipFile The zip file
      * @return A corpus object
      */
     public static Corpus fromTarball(File zipFile) {
+        return fromTarball(zipFile, null);
+    }
+
+    /**
+     * Create a corpus from a tarball (.tar.gz) file, each file will be
+     * considered a single document
+     *
+     * @param zipFile The zip file
+     * @param targetDir The directory to unzip files to
+     * @return A corpus object
+     */
+    public static Corpus fromTarball(File zipFile, File targetDir) {
         if (!zipFile.exists() && zipFile.isDirectory()) {
             throw new IllegalArgumentException(zipFile.getName() + " does not exist or is a folder");
         }
-        return new TarballCorpus(zipFile);
+        if (targetDir != null && !targetDir.mkdirs()) {
+            throw new IllegalArgumentException(targetDir.getName() + " could not be created as a file");
+        }
+        return new TarballCorpus(zipFile, targetDir);
     }
 
     private static class TarballCorpus implements Corpus {
 
         private final File zipFile;
+        private final File targetDir;
 
-        public TarballCorpus(File zipFile) {
+        public TarballCorpus(File zipFile, File targetDir) {
             this.zipFile = zipFile;
+            this.targetDir = targetDir;
         }
 
         @Override
@@ -249,18 +295,20 @@ public class CorpusTools {
                 public Iterator<Document> iterator() {
                     try {
                         final TarArchiveInputStream tais = new TarArchiveInputStream(new GzipCompressorInputStream(new FileInputStream(zipFile)));
-                        
+
                         return new Iterator<Document>() {
                             File file = null;
                             TarArchiveEntry tae;
-                            
+
                             private void advance() {
                                 try {
-                                    if(tae == null)
+                                    if (tae == null) {
                                         tae = tais.getNextTarEntry();
-                                    while(tae != null && !tae.isFile())
+                                    }
+                                    while (tae != null && !tae.isFile()) {
                                         tae = tais.getNextTarEntry();
-                                } catch(IOException x) {
+                                    }
+                                } catch (IOException x) {
                                     throw new RuntimeException(x);
                                 }
                             }
@@ -275,25 +323,35 @@ public class CorpusTools {
                             public Document next() {
                                 try {
                                     advance();
-                                    if(tae == null) throw new NoSuchElementException();
-                                    if (file != null) {
+                                    if (tae == null) {
+                                        throw new NoSuchElementException();
+                                    }
+                                    if (file != null && targetDir == null) {
                                         file.delete();
                                     }
-                                    file = File.createTempFile(tae.getName(), "");
-                                    file.deleteOnExit();
+                                    if (targetDir != null) {
+                                        file = new File(targetDir, tae.getName());
+                                        while (file.exists()) {
+                                            file = new File(targetDir, tae.getName() + new Random().nextInt(10000));
+                                        }
+                                    } else {
+                                        file = File.createTempFile(tae.getName(), "");
+                                        file.deleteOnExit();
+                                    }
                                     FileOutputStream fos2 = new FileOutputStream(file);
 
                                     fos2.getChannel().transferFrom(Channels.newChannel(tais), 0, Long.MAX_VALUE);
 
                                     return new Document(SaffronPath.fromFile(file), tae.getName(), null,
                                             tae.getName(), Files.probeContentType(new File(tae.getName()).toPath()),
-                                            new ArrayList<Author>(), new HashMap<String, String>(), null).withLoader(TIKA_LOADER);                                    
+                                            new ArrayList<Author>(), new HashMap<String, String>(), null).withLoader(TIKA_LOADER);
                                 } catch (IOException x) {
                                     throw new RuntimeException(x);
                                 } finally {
                                     try {
                                         tae = tais.getNextTarEntry();
-                                    } catch(IOException x) { }
+                                    } catch (IOException x) {
+                                    }
                                 }
                             }
                         };
